Se agrga buscador de usuarios

+++ TPC_Brandan_Repolledo/Dominio/Persona.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Persona
     {
-        public int DNI { get; set; }
+        public long DNI { get; set; }
 
         public string Nombre { get; set; }
